Extract FinderMethods to its own file and reorder included block
Move the growing FinderMethods module (find, find_by, find_by!,
exists?, and supporting private methods) into its own file at
app/models/work_package/semantic_identifier/finder_methods.rb.

Reorder the concern to place the included block at the top,
matching the convention used by other WorkPackage concerns
(Ancestors, Validations, Hooks, Journalized).
