+++ app/models/work_package/semantic_identifier.rb
@@ -36,97 +36,6 @@ module WorkPackage::SemanticIdentifier
   # The frontend equivalent lives in WP_ID_URL_PATTERN (work-package-id-pattern.ts).
   ID_ROUTE_CONSTRAINT = /\d+|[A-Za-z][A-Za-z0-9_]*-\d+/
 
-  # Finder methods that work on both the WorkPackage class and ActiveRecord::Relation scopes:
-  #   semantic_id?("PROJ-42") → true
-  #   semantic_id?(" 456 ")   → false  (stripped, then numeric)
-  #   semantic_id?("123")     → false
-  #   semantic_id?(123)       → false
-  module FinderMethods
-    def find(*args)
-      return find_by_id_or_identifier!(args.first.strip) if args.length == 1 && semantic_id?(args.first)
-
-      super
-    end
-
-    # Override find_by to transparently resolve semantic identifiers when called
-    # with `id:` as the sole keyword (e.g. `find_by(id: "PROJ-42")`).
-    # All other find_by calls pass through to ActiveRecord unchanged.
-    #
-    # AR's find_by signature is find_by(arg, *args) — it doesn't use keyword splat,
-    # so hash kwargs arrive as the positional `arg`. We match on that.
-    def find_by(*args)
-      if args.length == 1 && args.first.is_a?(Hash) && args.first.keys == [:id] && semantic_id?(args.first[:id])
-        find_by_id_or_identifier(args.first[:id])
-      else
-        super
-      end
-    end
-
-    # Mirror of find_by — Rails implements find_by! independently (not via find_by),
-    # so we must override both to keep the pair consistent.
-    def find_by!(*args)
-      if args.length == 1 && args.first.is_a?(Hash) && args.first.keys == [:id] && semantic_id?(args.first[:id])
-        find_by_id_or_identifier!(args.first[:id])
-      else
-        super
-      end
-    end
-
-    def exists?(conditions = :none)
-      return super unless semantic_id?(conditions)
-
-      exists_by_semantic_identifier?(conditions)
-    end
-
-    private
-
-    # Resolves any identifier form to a WorkPackage.
-    #   - Numeric string ("12345")    → find by primary key
-    #   - Semantic string ("PROJ-42") → lookup via work_packages table and alias table
-    #
-    # Returns nil on miss.
-    def find_by_id_or_identifier(identifier)
-      return find_by(id: identifier) unless semantic_id?(identifier)
-
-      find_by_semantic_identifier(identifier)
-    end
-
-    # Same as find_by_id_or_identifier but raises ActiveRecord::RecordNotFound on miss.
-    def find_by_id_or_identifier!(identifier)
-      find_by_id_or_identifier(identifier) ||
-        raise(ActiveRecord::RecordNotFound.new(
-                "Couldn't find WorkPackage with identifier=#{identifier}", "WorkPackage", "identifier", identifier
-              ))
-    end
-
-    # Returns true when value looks like a semantic work package identifier (e.g. "PROJ-42").
-    # Non-string values (Integer, Hash, nil, Array) and numeric strings ("123", " 456 ")
-    # return false — these fall through to standard ActiveRecord lookup.
-    def semantic_id?(value)
-      return false unless value.is_a?(String)
-
-      stripped = value.strip
-      stripped.to_i.to_s != stripped
-    end
-
-    # Looks up by current identifier column first, then falls back to
-    # the alias table for historical identifiers. Two-step because AR's
-    # .or() requires structurally compatible relations (joins breaks it).
-    def find_by_semantic_identifier(identifier)
-      find_by(identifier:) ||
-        by_semantic_alias(identifier).first
-    end
-
-    def exists_by_semantic_identifier?(identifier)
-      where(identifier:).exists? ||
-        by_semantic_alias(identifier).exists?
-    end
-
-    def by_semantic_alias(identifier)
-      joins(:semantic_aliases).where(work_package_semantic_aliases: { identifier: })
-    end
-  end
-
   included do
     has_many :semantic_aliases,
              class_name: "WorkPackageSemanticAlias",
